refactor ui into modular components
